refactor: rename h files to hpp for consistency
