+++ src/bundle.cpp
@@ -1,6 +1,6 @@
-#include "bundle.h"
+#include "bundle.hpp"
 
-#include "util.h"
+#include "util.hpp"
 
 #include <vector>
