+++ src/download.cpp
@@ -1,4 +1,4 @@
-#include "download.h"
+#include "download.hpp"
 
 #include <windows.h>
 #include <winhttp.h>
