+++ src/util.cpp
@@ -1,4 +1,4 @@
-#include "util.h"
+#include "util.hpp"
 
 #include <shlwapi.h>
 #include <windows.h>
